update form to use external hook to reduce code in component
